#221@minor: Move insertAdjacent functions to Element.
