+++ packages/happy-dom/src/nodes/node/INode.ts
@@ -2,7 +2,6 @@ import IEventTarget from '../../event/IEventTarget';
 import IDocument from '../document/IDocument';
 import IElement from '../element/IElement';
 
-export type TInsertAdjacentPositions = 'beforebegin' | 'afterbegin' | 'beforeend' | 'afterend';
 export default interface INode extends IEventTarget {
 	readonly ownerDocument: IDocument;
 	readonly parentNode: INode;
@@ -23,9 +22,6 @@ export default interface INode extends IEventTarget {
 	appendChild(node: INode): INode;
 	removeChild(node: INode): void;
 	insertBefore(newNode: INode, referenceNode?: INode | null): INode;
-	insertAdjacentElement(position: TInsertAdjacentPositions, node: INode): INode | null;
-	insertAdjacentHTML(position: TInsertAdjacentPositions, text: string): void;
-	insertAdjacentText(position: TInsertAdjacentPositions, text: string): void;
 	replaceChild(newChild: INode, oldChild: INode): INode;
 	toString(): string;
 }
